fix: satisfy static-check on the copy-segment inspector
Three golangci-lint violations introduced by 22ebd61cd:

- depguard: copy_segment_inspector.go imported the stdlib "errors";
  the repo mandates github.com/cockroachdb/errors. Only use site is
  errors.Is, which cockroachdb/errors provides identically.
- gosec G118: the WithCancel cancel func is stored as
  copySegmentInspector.dropCancel and invoked in Close(), so the
  warning is a false positive; annotated with the same //nolint:gosec
  form already used in internal/distributed/*/service.go.
- unconvert: storage.StorageV3 is already declared int64
  (internal/storage/rw.go:47), so the int64() conversion is redundant.

No behaviour change.

+++ internal/datacoord/snapshot_manager_test.go
@@ -1847,7 +1847,7 @@ func TestCreateRestoreJob_PreRegistersTargetSegmentsAsImporting(t *testing.T) {
 		assert.Equal(t, "dst-ch", seg.GetInsertChannel())
 		assert.Equal(t, commonpb.SegmentState_Importing, seg.GetState())
 		assert.True(t, seg.GetIsImporting())
-		assert.Equal(t, int64(storage.StorageV3), seg.GetStorageVersion())
+		assert.Equal(t, storage.StorageV3, seg.GetStorageVersion())
 		basePath, version, err := packed.UnmarshalManifestPath(seg.GetManifestPath())
 		require.NoError(t, err)
 		assert.Equal(t, "files/insert_log/200/20/2001", basePath)
